+++ lgn-messages/src/v1/preprocessing/mod.rs
@@ -1,6 +1,3 @@
-use db_tasks::BatchedIndex;
-use ext_tasks::BatchedContract;
-use ext_tasks::BatchedLength;
 use mp2_v1::api::MAX_FIELD_PER_EVM;
 use mp2_v1::values_extraction;
 use serde_derive::Deserialize;
@@ -9,8 +6,7 @@ use serde_derive::Serialize;
 use super::query::MAX_NUM_COLUMNS;
 use super::ConcreteCircuitInput;
 
-pub mod db_tasks;
-pub mod ext_tasks;
+pub mod batched;
 
 pub type ConcreteValueExtractionCircuitInput =
     values_extraction::CircuitInput<69, MAX_NUM_COLUMNS, MAX_FIELD_PER_EVM>;
@@ -18,8 +14,8 @@ pub type ConcreteValueExtractionCircuitInput =
 #[derive(Deserialize, Serialize)]
 #[serde(tag = "type")]
 pub enum PreprocessingTask {
-    BatchedIndex(BatchedIndex),
-    BatchedLength(BatchedLength),
-    BatchedContract(BatchedContract),
+    BatchedIndex(batched::BatchedIndex),
+    BatchedLength(batched::BatchedLength),
+    BatchedContract(batched::BatchedContract),
     CircuitInput(ConcreteCircuitInput),
 }
